Simplify traits
