+++ pallets/simple-nft/src/mock.rs
@@ -105,9 +105,9 @@ impl Default for ProcessIdentifier {
 pub struct MockProcessValidator {}
 
 type TestProcessId = ProcessFullyQualifiedId<ProcessIdentifier, u32>;
-type TestProcessIO = ProcessIO<ConstU32<2>, u64, Role, ConstU32<4>, u64, MetadataValue<u64>>;
+type TestProcessIO = ProcessIO<u64, Role, u64, MetadataValue<u64>>;
 
-impl ProcessValidator<ConstU32<2>, u64, Role, ConstU32<4>, u64, MetadataValue<u64>> for MockProcessValidator {
+impl ProcessValidator<u64, Role, u64, MetadataValue<u64>> for MockProcessValidator {
     type ProcessIdentifier = ProcessIdentifier;
     type ProcessVersion = u32;
